GitlabComment knows how to deal with project ids (#68)
`ListGitlabMergeRequests` only provides the project's id in the status,
as that's what the API returns.  `GitlabComment` only accepts the
project's path.

Not an excellent situation, but one we can readily handle.

For most of the GitLab API, it will accept either a string (the "full
path") or a numeric id to identify a project.  Here we simply change the
`Project` from a `string` to `intstr`, and inject some logic to handle
it.

Note that this does introduce a non-breaking CRD change.